+++ lib/jsx_rosetta/ir.rb
@@ -6,12 +6,12 @@ require_relative "ir/lowering"
 
 module JsxRosetta
   module IR
-    def self.lower(ast_file, source:)
-      Lowering.lower(ast_file, source: source)
+    def self.lower(ast_file, source:, keep_slot: false)
+      Lowering.lower(ast_file, source: source, keep_slot: keep_slot)
     end
 
-    def self.lower_all(ast_file, source:)
-      Lowering.lower_all(ast_file, source: source)
+    def self.lower_all(ast_file, source:, keep_slot: false)
+      Lowering.lower_all(ast_file, source: source, keep_slot: keep_slot)
     end
   end
 end